feat: add type converter to relay
