+++ src/HotChocolate/Core/src/Types/Types/Relay/Descriptors/NodeDescriptor~2.cs
@@ -61,6 +61,7 @@ namespace HotChocolate.Types.Relay.Descriptors
                 }
 
                 return null;
+
             });
         }
